Fixing bugs and moving existing tests (partway done)
